+++ app/controllers/meal_plan_recipes_controller.rb
@@ -1,5 +1,7 @@
 class MealPlanRecipesController < ApplicationController
   include KitchenHelpers
+  include MealCellResponder
+
   def create
     return head :unprocessable_entity if params[:recipe_id].blank?
 
@@ -17,61 +19,17 @@ class MealPlanRecipesController < ApplicationController
   end
 
   def destroy
-    mpr = MealPlanRecipe.find(params[:id])
+    mpr       = MealPlanRecipe.find(params[:id])
     meal_plan = mpr.meal_plan
+    date      = meal_plan.planned_on
+    slot      = meal_plan.meal_slot
     mpr.destroy
     meal_plan.reload
     if meal_plan.recipes.empty? && meal_plan.meal_plan_items.empty?
-      date = meal_plan.planned_on
-      slot = meal_plan.meal_slot
       meal_plan.destroy
       respond_with_empty_cell(date, slot)
     else
       respond_with_cell(meal_plan)
-    end
-  end
-
-  private
-
-  def respond_with_cell(meal_plan)
-    date = meal_plan.planned_on
-    slot = meal_plan.meal_slot
-    still_exists = MealPlan.exists?(meal_plan.id)
-
-    respond_to do |format|
-      format.turbo_stream do
-        render turbo_stream: turbo_stream.replace(
-          "meal_cell_#{date.iso8601}_#{slot}",
-          partial: "kitchen/meal_cell",
-          locals: {
-            date:               date,
-            slot:               slot,
-            meal_plan:          still_exists ? MealPlan.includes(:meal_plan_recipes, :recipes, meal_plan_items: :food_item).find(meal_plan.id) : nil,
-            recipes_for_slot:   Recipe.active.by_meal_type(slot).ordered,
-            food_items_grouped: grouped_food_items
-          }
-        )
-      end
-      format.html { redirect_to kitchen_path }
-    end
-  end
-
-  def respond_with_empty_cell(date, slot)
-    respond_to do |format|
-      format.turbo_stream do
-        render turbo_stream: turbo_stream.replace(
-          "meal_cell_#{date.iso8601}_#{slot}",
-          partial: "kitchen/meal_cell",
-          locals: {
-            date:               date,
-            slot:               slot,
-            meal_plan:          nil,
-            recipes_for_slot:   Recipe.active.by_meal_type(slot).ordered,
-            food_items_grouped: grouped_food_items
-          }
-        )
-      end
-      format.html { redirect_to kitchen_path }
     end
   end
 end